% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bm.R
\name{get.bm}
\alias{get.bm}
\title{generate a sample 2d Brownian motion (BM)}
\usage{
get.bm(time.seq, mu = 0, sigma2 = 1)
}
\arguments{
\item{time.seq}{input time sequence.}

\item{mu}{input expected value of BM or the starting value.}

\item{sigma2}{input variance}
}
\value{
x.t a series of numbers
}
\description{
generate a sample 2d Brownian motion (BM)
}
\examples{
get.bm(c(1,2,3,4))
}

