% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bm.R
\name{get_bm}
\alias{get_bm}
\title{generate a sample 2d Brownian motion (BM)}
\usage{
get_bm(time_seq, mu = 0, sigma2 = 1)
}
\arguments{
\item{time_seq}{input time sequence.}

\item{mu}{input expected value of BM or the starting value.}

\item{sigma2}{input variance}
}
\value{
x_t a series of numbers
}
\description{
generate a sample 2d Brownian motion (BM)
}
\examples{
get_bm(c(1,2,3,4))
}

