% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bm.R
\name{plot_bm}
\alias{plot_bm}
\title{plot a 2D Brownian motion (BM)}
\usage{
plot_bm(x, y, ...)
}
\arguments{
\item{x}{input x axis}

\item{y}{input y axis}

\item{...}{extra arguments to be passed to plot}
}
\value{
none
}
\description{
plot a 2D Brownian motion (BM)
}
\examples{
delta  = 1/32
t_seq  = seq(1, 100, delta)
sigma2 = 0.001
x_t    = get_bm(t_seq, sigma2 = sigma2)
y_t    = get_bm(t_seq, sigma2 = sigma2)
plot_bm(x_t, y_t)
}

